+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased changes
 
 - add `t:Floki.html_node/1` to possible `t:Pages.Html.html/0` types.
+- add `Pages.Driver.LiveView.live_redirect/2`
 
 ## 0.2.5